What I turned in
Score of 70/100. Comments: Alright so included in the CS1410 library there was a dialog box that you could use to ask for input and another to display the output, instead of having to create your own using JOptionPane. Also, the dialog boxes should return whatever is input so there is no need to use a scanner for input for this assignment. The output format that we wanted was found in the assignment because we were looking for the MPG along with others, not just the input that was given by the user. If you have any questions on the specs of an assignment don't hesitate to come to help hours or ask on piazza.